v0.0.2 - systems cleanup (#1)
* only host monitor route in non-prod mode

* move the schemas to the shared core-game package (also remove strictNullChecks=false)

* cleanup toast messages

* bump versions of apps and packages to v0.0.2

* refactor server room code into systems

* make enemies private on game room

* remove state view check

* cleanup Phaser scene imports of Phaser global

* add fpsDisplay and update pingDisplay

* update Enemy object

* update Player object

* create UISystem

* create InputSystem

* create RoomSystem

* create EnemySystem

* create RemotePlayerSystem

* add connection check on focus (to handle active connections to closed rooms)

* create PlayerSystem

* update game scene to use new systems

* add back game results GQL query

* cleanup type for game over scene (use GQL generated type)

* add back RESULTS logic (accidentally removed when refactoring)

* add back Coin Modal logic (accidentally removed when refactoring)

* cleanup devLog styles

* add enemy position interpolation

* rename dev log to proper number

* fix missing space

* add dev log 1 - Demo Cleanup (riveting...)

+++ apps/game-api/src/graphql/resolvers.ts
@@ -8,6 +8,7 @@ export const resolvers: Resolvers<Context> = {
       return true;
     },
     profile: async (_, __, { dataSources, user }) => {
+      if (!user) return null;
       return dataSources.profilesDb.getProfileByUserId(user.id);
     },
     userExists: async (_, { userName }, { dataSources }) => {
@@ -27,18 +28,21 @@ export const resolvers: Resolvers<Context> = {
   },
   Mutation: {
     createProfile: async (_, { userName }, { dataSources, user }) => {
+      if (!user) return null;
       return dataSources.profilesDb.createProfile({
         userId: user.id,
         userName,
       });
     },
     updateProfile: async (_, { userName }, { dataSources, user }) => {
+      if (!user) return null;
       return dataSources.profilesDb.updateProfile({
         userId: user.id,
         userName,
       });
     },
     deleteProfile: async (_, __, { authClient, dataSources, user }) => {
+      if (!user) return false;
       try {
         await dataSources.profilesDb.deleteProfile(user.id);
         await authClient.deleteUser(user.id);
